Remove TypeScript
This removes TypeScript and all dependencies and only keeps a simple
Javascript file that does not have any dependencies and can be run
directly. There is not need for an NPM build any more and we just
call a shell script via Javascript.

With this, we no longer have to commit node_modules and Javascript
compiled from TypeScript code. Also, the shell script can easily be run
directly for local testing.

Signed-off-by: Reinhard Naegele <unguiculus@gmail.com>