feat: add a "match_tags" option
I use this action to release the chart of [Mercure](https://mercure.rocks).
We use a monolithic Git repository containing several projects, including the Helm chart.
Currently, this action isn't compatible with our setup because it computes the changes
made to the chart since the previous tag.

In our case we also use tags for other projects in the same repository,
so this actions is never triggered because there are no changes with the "previous" tag (which usually is for another project in the same repository).

This PR introduces a `match_tags` option allowing to filter the tags to match using a glob.

Signed-off-by: Kévin Dunglas <dunglas@gmail.com>