Support actions triggered by a tag
The original helm/chart-releaser-action does not work on actions triggered by a tag because the `lookup_changed_charts` will never find differences in the charts. The problem is that `lookup_latest_tag` returns the tag that triggered the action, which is the current commit analyzed by `lookup_changed_charts`.

This simple solution skips the current (tagged) commit when searching for tags.

Signed-off-by: Luis Gracia <lgraval@gmail.com>