Write chart_version only if latest_tag is defined (#202)
The chart_version output was written regardless of whether the
latest_tag variable was defined. The latest_tag variable is only defined
if skip_packaging is not set. This leads to issue #171.

Setting skip_packaging implies that an external process takes care of
packaging the chart and setting the chart version. There is thus no need
to write the chart_version output as it should already be known.

Closes #171

Signed-off-by: Ferdinand Hofherr <mail@ferdinandhofherr.de>